Add pegs to DebugVisualization
